FIX: Refactor 2
