+++ miscellaneous/metaDataRemover/txDataFormater.go
@@ -8,6 +8,31 @@ import (
 
 const ESDTDeleteMetadataPrefix = "ESDTDeleteMetadata"
 
+func createShardTxsDataMap(shardTokensMap map[uint32]map[string]struct{}, tokensToDeletePerTx uint64) (map[uint32][][]byte, error) {
+	shardTxsDataMap := make(map[uint32][][]byte)
+	for shardID, tokens := range shardTokensMap {
+		log.Info("creating txs data", "shardID", shardID, "num tokens", len(tokens))
+		tokensSorted, err := sortTokensIDByNonce(tokens)
+		if err != nil {
+			return nil, err
+		}
+
+		tokensIntervals := groupTokensByIntervals(tokensSorted)
+		tokensSortedByNonces := sortTokenIntervalsByMaxConsecutiveNonces(tokensIntervals)
+		tokensInBulks := groupTokenIntervalsInBulks(tokensSortedByNonces, tokensToDeletePerTx)
+
+		txsData, err := createTxsData(tokensInBulks)
+		if err != nil {
+			return nil, err
+		}
+
+		log.Info("created", "num of txs", len(txsData), "shardID", shardID, "num of nonces per tx", tokensToDeletePerTx)
+		shardTxsDataMap[shardID] = txsData
+	}
+
+	return shardTxsDataMap, nil
+}
+
 func createTxsData(bulks [][]*tokenData) ([][]byte, error) {
 	txsData := make([][]byte, 0, len(bulks))
 	for _, bulk := range bulks {
